Update .gitignore, pre-commit configuration, and enhance Cargo files
- Updated .gitignore to exclude new directories and files related to Python probing extensions and added patterns for temporary files.
- Modified .pre-commit-config.yaml to enable `cargo-clippy` and `cargo-fmt` hooks for Rust code linting and formatting.
- Enhanced Cargo.toml and Cargo.lock to include new dependencies and features, improving the overall project structure.
- Introduced a new clippy.toml file for shared Clippy settings to standardize linting across the project.
- Updated documentation to reflect changes in the development setup and project structure.

+++ probing/extensions/gpu/src/extensions/backend/apple/mod.rs
@@ -97,7 +97,7 @@ impl GpuBackend for AppleSiliconBackend {
     fn sample_memory(&self, ordinal: i32) -> Option<GpuMemorySample> {
         let dev = self.devices.iter().find(|d| d.ordinal == ordinal)?;
         let perf = ioreg::read_performance_stats();
-        let allocated = dev.device.current_allocated_size() as u64;
+        let allocated = dev.device.current_allocated_size();
         let budget = dev.device.recommended_max_working_set_size();
         let total = if budget > 0 {
             budget
